+++ armory/scenarios/image_classification.py
@@ -5,7 +5,7 @@ General image recognition scenario for image classification and object detection
 import numpy as np
 
 from armory.scenarios.scenario import Scenario
-from armory.utils.export import ImageClassificationExporter
+from armory.instrument.export import ImageClassificationExporter
 
 
 class ImageClassificationTask(Scenario):
