+++ armory/scenarios/video_ucf101_scenario.py
@@ -5,7 +5,7 @@ Scenario Contributor: MITRE Corporation
 """
 
 from armory.scenarios.scenario import Scenario
-from armory.utils.export import VideoClassificationExporter
+from armory.instrument.export import VideoClassificationExporter
 
 
 class Ucf101(Scenario):
